+++ firmware/sigmadrive/velocityprofiler/trapezoidprofile.h
@@ -6,9 +6,9 @@
 #include <stdint.h>
 
 std::vector<std::vector<int64_t>> 
-CalculateTrapezoidPoints(float Xi, float Xf, float Vin, float Vfin, float Vmax, float Amax, float Dmax, float Hz);
+CalculateTrapezoidPoints(int64_t Xi, int64_t Xf, int64_t Vin, int64_t Vfin, int64_t Vmax, int64_t Amax, int64_t Dmax, int64_t Hz);
 
 std::vector<std::vector<std::vector<int64_t>>> 
-CalculateTrapezoidPointsXY(float Xin, float Yin, float Xfin, float Yfin, float Vmax, float Accel, float Decel, float Hz);
+CalculateTrapezoidPointsXY(int64_t Xin, int64_t Yin, int64_t Xfin, int64_t Yfin, int64_t Vmax, int64_t Accel, int64_t Decel, int64_t Hz);
 
 #endif
